Back-django/requirements.txt.txt
Django>2.1.5
django-cors-headers
djangorestframework
markdown
django-filter